Adjust chat UI positioning and status icon
Changes made in Response.tsx and Message.tsx under the src/components/ directory, mainly to improve the UI of the chat feature and to better reflect the online status of the users. By adding the "clearfix" class to the 'li' tag, we have fixed the floating issue of chat bubbles. Also, the position of username and time were tweaked to be more visually pleasing. Finally, the 'fa-circle' class was modified to 'fa-circle me' in Response.tsx to distinguish the status of the current user in a more meaningful way.